backend/requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.2
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.10.1
phonenumbers==8.12.57
platformdirs==2.5.2
pytz==2022.4
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
